# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Vowed API",
    "version": "0.1.0",
    "description": "Backend for the Vowed Android app. All endpoints except /v1/health, /v1/meta and /v1/auth/* need `Authorization: Bearer <token>` from /v1/auth/verify. u64 amounts are decimal strings. Transaction builders return unsigned base64 transactions: the app must decode and check them before asking the wallet to sign."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/v1/challenges": {
      "get": {
        "summary": "List challenges (mine=true: only those I joined or created)",
        "tags": [
          "challenges"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "Open",
                "Settling",
                "Settled",
                "Voided"
              ]
            },
            "in": "query",
            "name": "status",
            "required": false
          },
          {
            "schema": {
              "default": "false",
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "in": "query",
            "name": "mine",
            "required": false
          },
          {
            "schema": {
              "default": 50,
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            },
            "in": "query",
            "name": "limit",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "challenges": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "pool": {
                            "type": "string"
                          },
                          "creator": {
                            "type": "string"
                          },
                          "mint": {
                            "type": "string"
                          },
                          "vault": {
                            "type": "string"
                          },
                          "kind": {
                            "type": "string"
                          },
                          "mode": {
                            "type": "string"
                          },
                          "penaltyBps": {
                            "type": "number"
                          },
                          "feeBps": {
                            "type": "number"
                          },
                          "startTs": {
                            "type": "number"
                          },
                          "endTs": {
                            "type": "number"
                          },
                          "joinDeadlineTs": {
                            "type": "number"
                          },
                          "settleAfterTs": {
                            "type": "number"
                          },
                          "durationDays": {
                            "type": "number"
                          },
                          "requiredDays": {
                            "type": "number"
                          },
                          "goalHash": {
                            "type": "string"
                          },
                          "participantCount": {
                            "type": "number"
                          },
                          "maxParticipants": {
                            "type": "number"
                          },
                          "settledCount": {
                            "type": "number"
                          },
                          "pendingClaims": {
                            "type": "number"
                          },
                          "totalDeposits": {
                            "type": "string"
                          },
                          "totalForfeit": {
                            "type": "string"
                          },
                          "totalSuccessStake": {
                            "type": "string"
                          },
                          "distributable": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string"
                          },
                          "squadId": {
                            "nullable": true,
                            "type": "string"
                          },
                          "trustTier": {
                            "nullable": true,
                            "type": "string",
                            "enum": [
                              "high",
                              "medium",
                              "low"
                            ]
                          },
                          "plan": {
                            "nullable": true
                          }
                        },
                        "required": [
                          "pool",
                          "creator",
                          "mint",
                          "vault",
                          "kind",
                          "mode",
                          "penaltyBps",
                          "feeBps",
                          "startTs",
                          "endTs",
                          "joinDeadlineTs",
                          "settleAfterTs",
                          "durationDays",
                          "requiredDays",
                          "goalHash",
                          "participantCount",
                          "maxParticipants",
                          "settledCount",
                          "pendingClaims",
                          "totalDeposits",
                          "totalForfeit",
                          "totalSuccessStake",
                          "distributable",
                          "status",
                          "squadId",
                          "trustTier",
                          "plan"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "challenges"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/challenges/{pool}": {
      "get": {
        "summary": "One challenge with its participants and my claimable amount",
        "tags": [
          "challenges"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "pool",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "challenge": {
                      "type": "object",
                      "properties": {
                        "pool": {
                          "type": "string"
                        },
                        "creator": {
                          "type": "string"
                        },
                        "mint": {
                          "type": "string"
                        },
                        "vault": {
                          "type": "string"
                        },
                        "kind": {
                          "type": "string"
                        },
                        "mode": {
                          "type": "string"
                        },
                        "penaltyBps": {
                          "type": "number"
                        },
                        "feeBps": {
                          "type": "number"
                        },
                        "startTs": {
                          "type": "number"
                        },
                        "endTs": {
                          "type": "number"
                        },
                        "joinDeadlineTs": {
                          "type": "number"
                        },
                        "settleAfterTs": {
                          "type": "number"
                        },
                        "durationDays": {
                          "type": "number"
                        },
                        "requiredDays": {
                          "type": "number"
                        },
                        "goalHash": {
                          "type": "string"
                        },
                        "participantCount": {
                          "type": "number"
                        },
                        "maxParticipants": {
                          "type": "number"
                        },
                        "settledCount": {
                          "type": "number"
                        },
                        "pendingClaims": {
                          "type": "number"
                        },
                        "totalDeposits": {
                          "type": "string"
                        },
                        "totalForfeit": {
                          "type": "string"
                        },
                        "totalSuccessStake": {
                          "type": "string"
                        },
                        "distributable": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string"
                        },
                        "squadId": {
                          "nullable": true,
                          "type": "string"
                        },
                        "trustTier": {
                          "nullable": true,
                          "type": "string",
                          "enum": [
                            "high",
                            "medium",
                            "low"
                          ]
                        },
                        "plan": {
                          "nullable": true
                        }
                      },
                      "required": [
                        "pool",
                        "creator",
                        "mint",
                        "vault",
                        "kind",
                        "mode",
                        "penaltyBps",
                        "feeBps",
                        "startTs",
                        "endTs",
                        "joinDeadlineTs",
                        "settleAfterTs",
                        "durationDays",
                        "requiredDays",
                        "goalHash",
                        "participantCount",
                        "maxParticipants",
                        "settledCount",
                        "pendingClaims",
                        "totalDeposits",
                        "totalForfeit",
                        "totalSuccessStake",
                        "distributable",
                        "status",
                        "squadId",
                        "trustTier",
                        "plan"
                      ],
                      "additionalProperties": false
                    },
                    "participants": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "wallet": {
                            "type": "string"
                          },
                          "stake": {
                            "type": "string"
                          },
                          "tzOffsetMinutes": {
                            "type": "number"
                          },
                          "daysCompleted": {
                            "type": "number"
                          },
                          "checkinBitmap": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "wallet",
                          "stake",
                          "tzOffsetMinutes",
                          "daysCompleted",
                          "checkinBitmap",
                          "status"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "me": {
                      "type": "object",
                      "properties": {
                        "joined": {
                          "type": "boolean"
                        },
                        "claimable": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "joined",
                        "claimable"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "challenge",
                    "participants",
                    "me"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/challenges/sync": {
      "post": {
        "summary": "Tell the backend about a transaction you sent so the mirror updates immediately (it re-reads the chain; nothing is trusted from the request)",
        "tags": [
          "challenges"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "signature": {
                    "type": "string",
                    "minLength": 80,
                    "maxLength": 100
                  }
                },
                "required": [
                  "signature"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "events": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "events"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}